Restructured venue handlers.
