+++ BingeBuddyNg/BingeBuddyNg.Services/Ranking/Querys/GetDrinksRankingQuery.cs
@@ -1,11 +1,38 @@
-﻿using MediatR;
+﻿using BingeBuddyNg.Services.Statistics;
+using BingeBuddyNg.Services.User;
+using MediatR;
 using System;
 using System.Collections.Generic;
-using System.Text;
+using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
 
 namespace BingeBuddyNg.Services.Ranking.Querys
 {
     public class GetDrinksRankingQuery : IRequest<List<UserRankingDTO>>
     {
+    }
+
+    public class GetDrinksRankingQueryHandler :
+       IRequestHandler<GetDrinksRankingQuery, List<UserRankingDTO>>
+    {
+        private readonly IUserRepository userRepository;
+        private readonly IUserStatsRepository userStatsRepository;
+
+        public GetDrinksRankingQueryHandler(IUserRepository userRepository, IUserStatsRepository userStatsRepository)
+        {
+            this.userRepository = userRepository ?? throw new ArgumentNullException(nameof(userRepository));
+            this.userStatsRepository = userStatsRepository ?? throw new ArgumentNullException(nameof(userStatsRepository));
+        }
+
+        public async Task<List<UserRankingDTO>> Handle(GetDrinksRankingQuery request, CancellationToken cancellationToken)
+        {
+            var userStats = await this.userStatsRepository.GetRankingStatisticsAsync();
+            var userIds = userStats.Select(u => u.UserId).Distinct();
+            var users = await this.userRepository.GetUsersAsync(userIds);
+
+            var result = userStats.Select(s => new UserRankingDTO(users.Select(u => new UserInfoDTO(u.Id, u.Name)).First(u => u.UserId == s.UserId), s.ToDto())).ToList();
+            return result;
+        }
     }
 }
